Six hours! Six hours for one css media query attribute!
I also modified the viewport settings in the index.html file so that likely had
an effect also... something I should follow up on.

But yeah, for whatever reason, the orientation attribute was preventing the
mobile phone MediaQuery component from properly loading. This makes no sense to
me, as the minimum width was not changing, nor was the orientation. Perhaps the
requirement that the attribute be described as 'portrait' instead of
'primary-portrait' was a factor? I am at a loss. Either way, things worked out.
Somehow...