Refactor With Breaking Changes to Be Fixed Tomorrow.
I pulled the payment section into a different component and created a new
profile object to separately hold user information. Much like the order object,
the profile object will be packed into the final Fetch payload along with the
Stripe token. Everything will then get parsed and stored in the database, with a
nice little email sent to the customer and the salespeople informing them of the
purchase.

The react-router-dom's Switch component was giving me hell. I ended up just
storing all the routes in a div and things are behaving as I would like. Clearly
I'm missing something fundamental on that front.