Set up email utility
Created a simple python class for sending emails to Nick Patrick after the
charge and order values are validated. The email password is sitting in plain
sight, so I'll need to change it when production rolls around. I should take
that as a cue to set up a Vault server.