Modified Styles. Changed Order Entry Logic.
The entry components are now being placed in different areas depending on the
screen orientation. I changed the logic for rendering a button so that the user
can only pick one fruit at a time. However, I need to update it further so that
users may toggle between fruits instead of requiring them to submit or remove
the currently highlighted fruit.

The button grid is still acting funky. Need to determine a proper approach for
expanding content to fit the available area. The current approach sort of
'half-works'.