Pulled Balance Alert into Separate Container. Corrected Some Interfaces.
The user is now shown the balance after they update the order. When I added the
SVGs and modified the names of the fruits, I neglected to update the order
object and the orderform interface to reflect the changes. Consequently, pina
colada was not update to read pineapple and peach was not updated to read
coconut.