Finally figured out router timeouts. Worked on styles.
In a stroke of stupidity, I neglected to thoroughly read the
react-router documentation. Turns out... there was a significant
update somewhat recently that enables me to access url history, which
makes it signficantly easier to build a custom link that runs a
timeout function which will make room for transition animations.

On a different note. I discovered that the logo header was padded off
the left screen because I was setting the width too large. I'm
skeptical of my little SVGResizer function, but it will do for now.