Included Profile and Order Summaries in the Payment Form.
Found another benefit of Rust along the way. I assigned the value for
the stateful order object (as described by "this.props.store" in
PaymentForm.tsx) to a separate variable (i.e.: let orderCopy =
this.store.order).

Were I more experienced in JS, I would have understood how dangerous
this is, because I can now modify this.props.store indirectly through
the orderCopy object, which is precisely what I did!

This introduced a bug that would have been difficult to acheive in
Rust due to it's implementation of ownership, a feature that ensures
memory safety.

Note to self!