Refactor: Removed location reducer.
I initially had been using a location reducer to manage
animations. The idea was that I would employ timeout functions to
update the website's route a couple seconds after the "location" had
changed. During that interval of time, the user would observe an outro
animation.

That was a confusing and messy solution. I've stopped worrying about
outro animations for the time being. I might even be moving towards an
all-too-convenient philosophical opposition!

Additionally, I modified the styles for the OrderForm so that they're
more in line with the format that's arisen in the other pages.